MVP - works with server on localhost

+++ SyncPlayServer/grails-app/domain/SyncPlayServer/Client.groovy
@@ -3,5 +3,5 @@ package SyncPlayServer
 class Client {
 	String name
 	String movie
-	long seconds
+	double seconds
 }
